rename misc.ts to git.ts
https://coveord.atlassian.net/browse/CDX-320
