+++ packages/cli-e2e/__tests__/angular.specs.ts
@@ -15,7 +15,7 @@ import {EOL} from 'os';
 import {ProcessManager} from '../utils/processManager';
 import {Terminal} from '../utils/terminal/terminal';
 import {BrowserConsoleInterceptor} from '../utils/browserConsoleInterceptor';
-import {commitProject, undoCommit} from '../utils/misc';
+import {commitProject, undoCommit} from '../utils/git';
 
 describe('ui:create:angular', () => {
   let browser: Browser;
