+++ packages/cli-e2e/__tests__/vue.specs.ts
@@ -14,7 +14,7 @@ import {ProcessManager} from '../utils/processManager';
 import {deactivateEnvironmentFile, restoreEnvironmentFile} from '../utils/file';
 import {Terminal} from '../utils/terminal/terminal';
 import {BrowserConsoleInterceptor} from '../utils/browserConsoleInterceptor';
-import {commitProject, undoCommit} from '../utils/misc';
+import {commitProject, undoCommit} from '../utils/git';
 
 describe('ui:create:vue', () => {
   let browser: Browser;
